Move mock airtable into tests dir
Reduce the size of the package, and put test fixtures where they should be
