+++ tests/test_models.py
@@ -5,7 +5,7 @@ from django.test import TestCase
 from tests.models import Advert
 from wagtail_airtable.mixins import AirtableMixin
 from unittest.mock import patch
-from wagtail_airtable.tests import get_mock_airtable
+from .mock_airtable import get_mock_airtable
 
 
 class TestAirtableModel(TestCase):
