+++ tests/test_views.py
@@ -4,7 +4,7 @@ from django.urls import reverse
 
 from tests.models import Advert
 from unittest.mock import patch
-from wagtail_airtable.tests import get_mock_airtable
+from .mock_airtable import get_mock_airtable
 
 
 class TestAdminViews(TestCase):
